create author write book query, refactored helpfull functions to move Utility
